+++ src/components/index.ts
@@ -1,4 +1,4 @@
 export { default as SearchBox } from './search/SearchBox';
-export { default as Banner } from './banner/banner';
+export { default as Banner } from './Banner';
 export { default as MatjibListTags } from './MatjibList/Tags';
 export { default as Footer } from './Footer';
