+++ src/components/search/SearchKeyword.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { css, jsx } from '@emotion/react';
-import glassSolid from '../../assets/icons/glass-solid.svg';
+import glassSolid from '@/assets/icons/glass-solid.svg';
 
 const keywordStyle = css`
   margin-top: 10px;
